MDL-64935 course: activity_navigation.php
The `Jump To` button in boost was positioned on the same line as the Activity
and Resouces links and the containers which hold the anchor tags were not
handled for overflowing. In general, this is not bad as long as the Activity/
Resource links are short. But when the names of these links are long, or when
the screen width is narrow, they overlap with the jump button. To fix this,
CSS white-space property is helpful, when set to pre-wrap, it causes the text
inside the anchor tag to act like if there was word-wrap set on it.